move no_standalone_expect rule to dir and migrate existing logic and tests
