+++ crates/oxc_linter/src/rules.rs
@@ -520,7 +520,6 @@ mod promise {
 mod vitest {
     pub mod no_conditional_tests;
     pub mod no_import_node_test;
-    pub mod no_standalone_expect;
     pub mod prefer_to_be_falsy;
     pub mod prefer_to_be_object;
     pub mod prefer_to_be_truthy;
@@ -1008,7 +1007,6 @@ oxc_macros::declare_all_lint_rules! {
     unicorn::switch_case_braces,
     unicorn::text_encoding_identifier_case,
     unicorn::throw_new_error,
-    vitest::no_standalone_expect,
     vitest::no_conditional_tests,
     vitest::no_import_node_test,
     vitest::prefer_to_be_falsy,
