+++ crates/oxc_linter/src/rules/jest/no_standalone_expect/tests/mod.rs
@@ -1,0 +1,4 @@
+mod jest;
+mod vitest;
+
+use super::NoStandaloneExpect;
