+++ crates/oxc_linter/src/rules/jest/no_standalone_expect/tests/vitest.rs
@@ -1,51 +1,7 @@
-use oxc_diagnostics::OxcDiagnostic;
-use oxc_macros::declare_oxc_lint;
-use oxc_span::Span;
-
-use crate::{
-    context::LintContext,
-    fixer::{RuleFix, RuleFixer},
-    rule::Rule,
-    AstNode,
-};
-
-fn no_standalone_expect_diagnostic(span: Span) -> OxcDiagnostic {
-    // See <https://oxc.rs/docs/contribute/linter/adding-rules.html#diagnostics> for details
-    OxcDiagnostic::warn("Should be an imperative statement about what is wrong")
-        .with_help("Should be a command-like statement that tells the user how to fix the issue")
-        .with_label(span)
-}
-
-#[derive(Debug, Default, Clone)]
-pub struct NoStandaloneExpect;
-
-declare_oxc_lint!(
-    /// ### What it does
-    ///
-    /// Prevents `expect` statements outside of a `test` or `it` block. An `expect`
-    /// within a helper function (but outside of a `test` or `it` block) will not
-    /// trigger this rule.
-    ///
-    /// Statements like `expect.hasAssertions()` will NOT trigger this rule since these
-    /// calls will execute if they are not in a test block.
-    ///
-    /// ### Example
-    /// ```javascript
-    /// describe('a test', () => {
-    ///     expect(1).toBe(1);
-    /// });
-    /// ```
-    NoStandaloneExpect,
-    vitest,
-    correctness,
-);
-
-impl Rule for NoStandaloneExpect {
-    fn run<'a>(&self, node: &AstNode<'a>, ctx: &LintContext<'a>) {}
-}
-
 #[test]
 fn test() {
+    use super::PreferLowercaseTitle;
+    use crate::rule::RuleMeta;
     use crate::tester::Tester;
 
     let pass = vec![
